done models, factories, some routes: /subjects_id /subjects /professors

+++ database/migrations/2022_01_25_093542_create_professors_table.php
@@ -17,8 +17,8 @@ class CreateProfessorsTable extends Migration
             $table->id()->comment("ID преподавателя");
             $table->unsignedBigInteger("position_id")->comment("ID должности");
             $table->integer("personnel_number")->comment("Табельный номер");
-            $table->integer("ITN")->comment("ИНН");
-            $table->integer("INILA")->comment("СНИЛС");
+            $table->bigInteger("ITN")->comment("ИНН");
+            $table->bigInteger("INILA")->comment("СНИЛС");
             $table->string("department")->comment("Кафедра");
             $table->date("date_of_commencement_of_teaching_activity")->comment("Дата начала преподавательской деятельности");
             $table->string("description")->nullable(true)->comment("Описание");
